tasks

+++ Models/PersonsTasks.cs
@@ -7,7 +7,7 @@ namespace PersonsMVC.Models;
 public partial class PersonsTasks
 {
     [Key]
-    public int Idtask { get; set; }
+    public long Idtask { get; set; }
 
     public string? Description { get; set; }
